+++ kotori/io/protocol/util.py
@@ -4,7 +4,7 @@ import math
 import arrow
 import datetime
 from six import text_type
-from dateutil.tz import gettz
+from dateutil.tz import gettz, tzutc
 from dateutil.parser import parse
 from pyramid.settings import asbool
 from twisted.web import http
@@ -177,23 +177,27 @@ def slugify_datettime(dstring):
     return arrow.get(dstring).to('utc').format('YYYYMMDDTHHmmss')
 
 
-def parse_timestamp(timestamp):
+def parse_timestamp(timestamp, use_utc=False):
 
     if isinstance(timestamp, text_type):
 
         # FIXME: Maybe use system timezone here by default.
         # TODO: Make configurable via channel settings.
 
-        # HACK: Assume CET (Europe/Berlin) for human readable timestamp w/o timezone offset
+        # Handle naive human-readable timestamps w/o timezone offset.
         qualified = any([token in timestamp for token in ['Z', '+', ' CET', ' CEST']])
         if not qualified:
-            timestamp += ' CET'
+            # Assume CET (Europe/Berlin) (default), or UTC.
+            if use_utc:
+                timestamp += ' UTC'
+            else:
+                timestamp += ' CET'
 
         # Parse datetime string
         # Remark: Maybe use pandas.tseries.tools.parse_time_string?
         # TODO: Cache results of call to gettz to improve performance
         berlin = gettz('Europe/Berlin')
-        tzinfos = {'CET': berlin, 'CEST': berlin}
+        tzinfos = {'UTC': tzutc(), 'CET': berlin, 'CEST': berlin}
         timestamp = parse(timestamp, tzinfos=tzinfos)
 
     return timestamp