+++ CHANGES.rst
@@ -9,6 +9,11 @@ in progress
 - CI: Update to Grafana 9.3.0
 - DAQ: Mask ``PASSKEY`` variable coming from HTTP, emitted by Ecowitt
 
+Breaking changes
+----------------
+
+- DAQ: Obtain naive timestamps as UTC by default, instead of CET
+
 
 .. _kotori-0.27.0: